+++ scripts/Python/ESI/__init__.py
@@ -1,2 +1,3 @@
 from ESI.ESI_Handler import *
 from ESI.ESI_Auth_Handler import *
+from ESI.Neucore_Auth_Handler import *